requirements.txt
tensorflow==2.3.1
discord=2.3.1
request=2.31.0
boto3=1.29.0